refactory: 기타 파일 정리
