+++ backend/src/main/java/toonpick/app/config/DataDBConfig.java
@@ -16,7 +16,7 @@ import java.util.HashMap;
 
 @Configuration
 @EnableJpaRepositories(
-        basePackages = "toonpick.app.repository",
+        basePackages = "toonpick.app",
         entityManagerFactoryRef = "dataEntityManager",
         transactionManagerRef = "dataTransactionManager"
 )
