--- backend/src/main/java/toonpick/app/controller/AdminController.java
@@ -1,15 +1,0 @@
-package toonpick.app.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
-
-@Controller
-@ResponseBody
-public class AdminController {
-
-    @GetMapping("/admin")
-    public String adminPage(){
-        return "admin page";
-    }
-}
