--- backend/src/main/java/toonpick/app/controller/MainController.java
@@ -1,16 +1,0 @@
-package toonpick.app.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
-
-@Controller
-@ResponseBody
-public class MainController {
-
-    @GetMapping("/")
-    public String mainPage(){
-        return "main page";
-    }
-
-}
