+++ backend/src/main/java/toonpick/app/member/entity/Member.java
@@ -4,7 +4,7 @@ import jakarta.persistence.*;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import toonpick.app.entity.BaseTimeEntity;
+import toonpick.app.util.BaseTimeEntity;
 
 @Entity
 @Getter
