+++ backend/src/main/java/toonpick/app/review/service/WebtoonReviewService.java
@@ -9,14 +9,14 @@ import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.transaction.annotation.Transactional;
-import toonpick.app.dto.PagedResponseDTO;
+import toonpick.app.webtoon.dto.PagedResponseDTO;
 import toonpick.app.review.dto.WebtoonReviewCreateDTO;
 import toonpick.app.review.dto.WebtoonReviewDTO;
 import toonpick.app.review.entity.ReviewLike;
 import toonpick.app.member.entity.Member;
 import toonpick.app.webtoon.entity.Webtoon;
 import toonpick.app.review.entity.WebtoonReview;
-import toonpick.app.exception.ResourceNotFoundException;
+import toonpick.app.util.exception.ResourceNotFoundException;
 import toonpick.app.review.mapper.WebtoonReviewMapper;
 import toonpick.app.review.repository.ReviewLikeRepository;
 import toonpick.app.member.repository.MemberRepository;
