+++ backend/src/main/java/toonpick/app/util/exception/ResourceNotFoundException.java
@@ -1,4 +1,4 @@
-package toonpick.app.exception;
+package toonpick.app.util.exception;
 
 public class ResourceNotFoundException extends RuntimeException {
     public ResourceNotFoundException(String message) {
