+++ backend/src/main/java/toonpick/app/webtoon/controller/WebtoonDataUpdateController.java
@@ -1,4 +1,4 @@
-package toonpick.app.controller;
+package toonpick.app.webtoon.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
