+++ backend/src/main/java/toonpick/app/webtoon/dto/WebtoonUpdateRequestDTO.java
@@ -5,8 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import toonpick.app.entity.enums.AgeRating;
-import toonpick.app.entity.enums.SerializationStatus;
+import toonpick.app.webtoon.entity.enums.AgeRating;
+import toonpick.app.webtoon.entity.enums.SerializationStatus;
 
 import java.time.DayOfWeek;
 import java.time.LocalDate;
