+++ backend/src/main/java/toonpick/app/webtoon/entity/Webtoon.java
@@ -20,9 +20,9 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import toonpick.app.review.entity.WebtoonReview;
-import toonpick.app.entity.enums.AgeRating;
-import toonpick.app.entity.enums.Platform;
-import toonpick.app.entity.enums.SerializationStatus;
+import toonpick.app.webtoon.entity.enums.AgeRating;
+import toonpick.app.webtoon.entity.enums.Platform;
+import toonpick.app.webtoon.entity.enums.SerializationStatus;
 
 import java.time.DayOfWeek;
 import java.time.LocalDate;
