+++ backend/src/main/java/toonpick/app/webtoon/entity/enums/AgeRating.java
@@ -1,4 +1,4 @@
-package toonpick.app.entity.enums;
+package toonpick.app.webtoon.entity.enums;
 
 public enum AgeRating {
     ALL, AGE_12, AGE_15, AGE_19
