+++ backend/src/main/java/toonpick/app/webtoon/entity/enums/Platform.java
@@ -1,4 +1,4 @@
-package toonpick.app.entity.enums;
+package toonpick.app.webtoon.entity.enums;
 
 public enum Platform {
     NAVER, KAKAO, RIDIBOOKS, LEZHIN, TOOMICS, MRBLUE, BUFFTOON
