+++ backend/src/main/java/toonpick/app/webtoon/repository/WebtoonRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.transaction.annotation.Transactional;
 import toonpick.app.webtoon.entity.Webtoon;
-import toonpick.app.entity.enums.SerializationStatus;
+import toonpick.app.webtoon.entity.enums.SerializationStatus;
 
 import java.time.DayOfWeek;
 import java.util.List;
