+++ backend/src/main/java/toonpick/app/webtoon/service/WebtoonService.java
@@ -15,8 +15,8 @@ import toonpick.app.webtoon.dto.WebtoonUpdateRequestDTO;
 import toonpick.app.webtoon.entity.Author;
 import toonpick.app.webtoon.entity.Genre;
 import toonpick.app.webtoon.entity.Webtoon;
-import toonpick.app.entity.enums.SerializationStatus;
-import toonpick.app.exception.ResourceNotFoundException;
+import toonpick.app.webtoon.entity.enums.SerializationStatus;
+import toonpick.app.util.exception.ResourceNotFoundException;
 import toonpick.app.webtoon.mapper.WebtoonMapper;
 import toonpick.app.webtoon.repository.AuthorRepository;
 import toonpick.app.webtoon.repository.GenreRepository;
