+++ backend/src/test/java/toonpick/app/unit/service/WebtoonReviewServiceTest.java
@@ -10,9 +10,9 @@ import toonpick.app.review.dto.WebtoonReviewDTO;
 import toonpick.app.member.entity.Member;
 import toonpick.app.webtoon.entity.Webtoon;
 import toonpick.app.review.entity.WebtoonReview;
-import toonpick.app.entity.enums.Platform;
-import toonpick.app.entity.enums.SerializationStatus;
-import toonpick.app.exception.ResourceNotFoundException;
+import toonpick.app.webtoon.entity.enums.Platform;
+import toonpick.app.webtoon.entity.enums.SerializationStatus;
+import toonpick.app.util.exception.ResourceNotFoundException;
 import toonpick.app.member.repository.MemberRepository;
 import toonpick.app.webtoon.repository.WebtoonRepository;
 import toonpick.app.review.repository.WebtoonReviewRepository;
